+++ craft_platforms/charm/_build.py
@@ -40,7 +40,7 @@ If no platforms are defined, the charm will be built on and for these architectu
 """
 
 
-def _validate_base_definition(
+def _validate_base_definition(  # noqa: PLR0912
     base: Optional[str],
     build_base: Optional[str],
     platform_name: Optional[str],
@@ -68,16 +68,24 @@ def _validate_base_definition(
 
     if platform:
         if platform_base:
-            raise _errors.InvalidMultiBaseError(
-                message=(
-                    f"Platform {platform_name!r} declares a base in the platform's "
-                    "name and declares 'build-on' and 'build-for' entries."
-                ),
-                resolution=(
-                    "Either remove the base from the platform's name or remove the "
-                    "'build-on' and 'build-for' entries for the platform."
-                ),
-            )
+            build_for_bases = [
+                _architectures.parse_base_and_architecture(target_base)[0]
+                for target_base in platform.get("build-for", [platform_name])
+            ]
+            for check_base in build_for_bases:
+                if check_base in (None, platform_base):
+                    continue
+                raise _errors.InvalidMultiBaseError(
+                    message=(
+                        f"Platform {platform_name!r} declares a base in the "
+                        "platform's name and declares an incompatible 'build-for' "
+                        f"entry ({check_base})."
+                    ),
+                    resolution=(
+                        "Either remove the base from the platform's name or remove "
+                        "the incompatible 'build-for' entry for the platform."
+                    ),
+                )
         # create a set of the bases defined in the build-on and build-for entries
         bases = set()
         for entry in [